extract determineTransactionType from tx controller
