+++ app/scripts/controllers/transactions/index.test.js
@@ -1305,156 +1305,6 @@ describe('Transaction Controller', function () {
     });
   });
 
-  describe('#_determineTransactionType', function () {
-    it('should return a simple send type when to is truthy but data is falsy', async function () {
-      const result = await txController._determineTransactionType({
-        to: '0xabc',
-        data: '',
-      });
-      assert.deepEqual(result, {
-        type: TRANSACTION_TYPES.SIMPLE_SEND,
-        getCodeResponse: null,
-      });
-    });
-
-    it('should return a token transfer type when data is for the respective method call', async function () {
-      const result = await txController._determineTransactionType({
-        to: '0xabc',
-        data:
-          '0xa9059cbb0000000000000000000000002f318C334780961FB129D2a6c30D0763d9a5C970000000000000000000000000000000000000000000000000000000000000000a',
-      });
-      assert.deepEqual(result, {
-        type: TRANSACTION_TYPES.TOKEN_METHOD_TRANSFER,
-        getCodeResponse: undefined,
-      });
-    });
-
-    it('should return a token approve type when data is for the respective method call', async function () {
-      const result = await txController._determineTransactionType({
-        to: '0xabc',
-        data:
-          '0x095ea7b30000000000000000000000002f318C334780961FB129D2a6c30D0763d9a5C9700000000000000000000000000000000000000000000000000000000000000005',
-      });
-      assert.deepEqual(result, {
-        type: TRANSACTION_TYPES.TOKEN_METHOD_APPROVE,
-        getCodeResponse: undefined,
-      });
-    });
-
-    it('should return a contract deployment type when to is falsy and there is data', async function () {
-      const result = await txController._determineTransactionType({
-        to: '',
-        data: '0xabd',
-      });
-      assert.deepEqual(result, {
-        type: TRANSACTION_TYPES.DEPLOY_CONTRACT,
-        getCodeResponse: undefined,
-      });
-    });
-
-    it('should return a simple send type with a 0x getCodeResponse when there is data and but the to address is not a contract address', async function () {
-      const result = await txController._determineTransactionType({
-        to: '0x9e673399f795D01116e9A8B2dD2F156705131ee9',
-        data: '0xabd',
-      });
-      assert.deepEqual(result, {
-        type: TRANSACTION_TYPES.SIMPLE_SEND,
-        getCodeResponse: '0x',
-      });
-    });
-
-    it('should return a simple send type with a null getCodeResponse when to is truthy and there is data and but getCode returns an error', async function () {
-      const result = await txController._determineTransactionType({
-        to: '0xabc',
-        data: '0xabd',
-      });
-      assert.deepEqual(result, {
-        type: TRANSACTION_TYPES.SIMPLE_SEND,
-        getCodeResponse: null,
-      });
-    });
-
-    it('should return a contract interaction type with the correct getCodeResponse when to is truthy and there is data and it is not a token transaction', async function () {
-      const _providerResultStub = {
-        // 1 gwei
-        eth_gasPrice: '0x0de0b6b3a7640000',
-        // by default, all accounts are external accounts (not contracts)
-        eth_getCode: '0xa',
-      };
-      const _provider = createTestProviderTools({
-        scaffold: _providerResultStub,
-      }).provider;
-      const _fromAccount = getTestAccounts()[0];
-      const _blockTrackerStub = new EventEmitter();
-      _blockTrackerStub.getCurrentBlock = noop;
-      _blockTrackerStub.getLatestBlock = noop;
-      const _txController = new TransactionController({
-        provider: _provider,
-        getGasPrice() {
-          return '0xee6b2800';
-        },
-        networkStore: new ObservableStore(currentNetworkId),
-        getCurrentChainId: () => currentChainId,
-        txHistoryLimit: 10,
-        blockTracker: _blockTrackerStub,
-        signTransaction: (ethTx) =>
-          new Promise((resolve) => {
-            ethTx.sign(_fromAccount.key);
-            resolve();
-          }),
-        getParticipateInMetrics: () => false,
-      });
-      const result = await _txController._determineTransactionType({
-        to: '0x9e673399f795D01116e9A8B2dD2F156705131ee9',
-        data: 'abd',
-      });
-      assert.deepEqual(result, {
-        type: TRANSACTION_TYPES.CONTRACT_INTERACTION,
-        getCodeResponse: '0x0a',
-      });
-    });
-
-    it('should return a contract interaction type with the correct getCodeResponse when to is a contract address and data is falsy', async function () {
-      const _providerResultStub = {
-        // 1 gwei
-        eth_gasPrice: '0x0de0b6b3a7640000',
-        // by default, all accounts are external accounts (not contracts)
-        eth_getCode: '0xa',
-      };
-      const _provider = createTestProviderTools({
-        scaffold: _providerResultStub,
-      }).provider;
-      const _fromAccount = getTestAccounts()[0];
-      const _blockTrackerStub = new EventEmitter();
-      _blockTrackerStub.getCurrentBlock = noop;
-      _blockTrackerStub.getLatestBlock = noop;
-      const _txController = new TransactionController({
-        provider: _provider,
-        getGasPrice() {
-          return '0xee6b2800';
-        },
-        networkStore: new ObservableStore(currentNetworkId),
-        getCurrentChainId: () => currentChainId,
-        txHistoryLimit: 10,
-        blockTracker: _blockTrackerStub,
-        signTransaction: (ethTx) =>
-          new Promise((resolve) => {
-            ethTx.sign(_fromAccount.key);
-            resolve();
-          }),
-        getParticipateInMetrics: () => false,
-      });
-      const result = await _txController._determineTransactionType({
-        to: '0x9e673399f795D01116e9A8B2dD2F156705131ee9',
-        data: '',
-      });
-      assert.deepEqual(result, {
-        type: TRANSACTION_TYPES.CONTRACT_INTERACTION,
-        getCodeResponse: '0x0a',
-      });
-    });
-  });
-
   describe('#getPendingTransactions', function () {
     it('should show only submitted and approved transactions as pending transaction', function () {
       txController.txStateManager._addTransactionsToState([
