+++ ui/helpers/utils/util.js
@@ -65,14 +65,6 @@ export function isDefaultMetaMaskChain(chainId) {
   return false;
 }
 
-// Both inputs should be strings. This method is currently used to compare tokenAddress hex strings.
-export function isEqualCaseInsensitive(value1, value2) {
-  if (typeof value1 !== 'string' || typeof value2 !== 'string') {
-    return false;
-  }
-  return value1.toLowerCase() === value2.toLowerCase();
-}
-
 export function valuesFor(obj) {
   if (!obj) {
     return [];
