+++ ui/hooks/useCurrentAsset.js
@@ -3,11 +3,11 @@ import { useRouteMatch } from 'react-router-dom';
 import { getTokens } from '../ducks/metamask/metamask';
 import { getCurrentChainId } from '../selectors';
 import { ASSET_ROUTE } from '../helpers/constants/routes';
-import { isEqualCaseInsensitive } from '../helpers/utils/util';
 import {
   SWAPS_CHAINID_DEFAULT_TOKEN_MAP,
   ETH_SWAPS_TOKEN_OBJECT,
 } from '../../shared/constants/swaps';
+import { isEqualCaseInsensitive } from '../../shared/modules/string-utils';
 
 /**
  * Returns a token object for the asset that is currently being viewed.
