+++ ui/pages/settings/settings-search/settings-search.js
@@ -3,9 +3,9 @@ import PropTypes from 'prop-types';
 import Fuse from 'fuse.js';
 import InputAdornment from '@material-ui/core/InputAdornment';
 import TextField from '../../../components/ui/text-field';
-import { isEqualCaseInsensitive } from '../../../helpers/utils/util';
 import { I18nContext } from '../../../contexts/i18n';
 import SearchIcon from '../../../components/ui/search-icon';
+import { isEqualCaseInsensitive } from '../../../../shared/modules/string-utils';
 
 export default function SettingsSearch({
   onSearch,
